Simplify version handling of generated RBIs

+++ lib/rbi/cli_helper.rb
@@ -81,22 +81,17 @@ module RBI
           exit(1)
         end
       end
-      begin
-        gem_rbi_path = "#{ctx.path}/sorbet/rbi/gems"
-        if git
-          file_path = Dir["#{gem_rbi_path}/#{name}@#{version}-*.rbi"]
-          FileUtils.mv(file_path, ".")
-          logger.success("Generated RBI for `#{name}@#{version}`")
-        else
-          FileUtils.mv("#{gem_rbi_path}/#{name}@#{version}.rbi", ".")
-          logger.success("Generated `#{name}@#{version}.rbi`")
-        end
-      rescue Errno::ENOENT
-        logger.error(<<~ERR)
-          Unable to move gem RBI to target directory. Generated RBI must have a different version number than what you specifified.
-          Ensure version number to `rbi generate` command matches the version number retrieved from your specific source.
-        ERR
+      gem_rbi_path = "#{ctx.path}/sorbet/rbi/gems/#{name}@#{version}*.rbi"
+      files = Dir[gem_rbi_path]
+      if files.empty?
+        logger.error("Unable to generate RBI: no file matching #{gem_rbi_path}")
+        exit(1)
       end
+
+      file_path = T.must(files.first)
+      version_string = file_path.sub(/^.*@/, "").sub(/\.rbi$/, "")
+      FileUtils.mv(file_path, ".")
+      logger.success("Generated `#{name}@#{version_string}.rbi`")
 
       ctx.destroy
     end
